Add TocTree adapter
