+++ sphinx/environment/adapters/__init__.py
@@ -1,0 +1,10 @@
+# -*- coding: utf-8 -*-
+"""
+    sphinx.environment.adapters
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    Sphinx environment adapters
+
+    :copyright: Copyright 2007-2016 by the Sphinx team, see AUTHORS.
+    :license: BSD, see LICENSE for details.
+"""
